backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
aiohttp==3.7.4
python-can==3.3.4
pydantic==1.10.7